Further unit test adjustments to enure 100% correct test executions
